convert Rates service to TS
